+++ sdk/client/typescript/src/novellia/index.ts
@@ -1,0 +1,38 @@
+export {
+  NovelliaService,
+} from './novellia.service';
+export { 
+  GetProductsRequest,
+} from './_generated/apis';
+export {
+  Attribution,
+  CardanoTip,
+  CardanoTransaction,
+  Commission,
+  DescriptionSet,
+  Extended721Token,
+  MinterInfo,
+  NativeToken,
+  NovelliaProduct,
+  NovelliaStandardToken,
+  OccultaNovelliaCharacter,
+  OccultaNovelliaCharacterCard,
+  OccultaNovelliaCharacterProgression,
+  OccultaNovelliaCharacterStats,
+  OffChainResource,
+  Product,
+  ProductMarket,
+  ProductMetadata,
+  ProductOrganization,
+  ProductPricing,
+  ProductProduct,
+  ProductsList,
+  ProductStock,
+  Status as NovelliaStatus,
+  StatusCardano as NovelliaStatusCardano,
+  Token,
+  WorkflowInformation,
+} from './_generated/models';
+export {
+  ConfigurationParameters as NovelliaConfigurationParameters
+} from './_generated';