+++ sdk/client/typescript/src/order_fulfillment/_generated/models/index.ts
@@ -1,0 +1,6 @@
+export * from './Order';
+export * from './OrderCreated';
+export * from './OrderCustomer';
+export * from './OrderItems';
+export * from './OrderPayment';
+export * from './Status';