+++ sdk/client/typescript/src/order_fulfillment/index.ts
@@ -1,0 +1,14 @@
+export {
+  OrderFulfillmentService
+} from './order_fulfillment.service';
+export {
+  Order,
+  OrderCreated,
+  OrderCustomer,
+  OrderItems,
+  OrderPayment,
+  Status as OrderFulfillmentStatus,
+} from './_generated/models';
+export {
+  ConfigurationParameters as OrderFulfillmentConfigurationParameters
+} from './_generated';